requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
chromadb==0.5.3
langchain==0.2.7
langchain-community==0.2.7
sentence-transformers==3.0.1
python-dotenv==1.0.1
pydantic==2.8.2
pdfplumber==0.11.4
Pillow==10.4.0
pytesseract==0.3.10
numpy>=1.26.0
